BACKEND all Models

+++ backend/src/main/java/org/example/backend/recipe/model/NutritionValuesModel.java
@@ -3,9 +3,9 @@ package org.example.backend.recipe.model;
 import com.mongodb.lang.Nullable;
 
 public record NutritionValuesModel(
-        int fat,
-        int protein,
-        int carbs,
-        int calories,
+        double fat,
+        double protein,
+        double carbs,
+        double calories,
         @Nullable Double skaldeman ) {
 }
